Yet another local storage refactoring.
It separates the concept of "storage" (in practice, the sqlite3 tables)
from the concept of "local database" (i.e. the actual sqlite3 connection).
The local database is still called a BaseStorage here and will be renamed
to "LocalDatabase" in a later commit.

This commit also changes how the tables and databases are organized.
In particular, the user storage uses:
- the user local database to store the user manifest.
And the workspace storage uses:
- the cache local database to store the blocks
- the data local database to store the chunks and manifests.