Add postponed chunk cleanup
Before this change, there was a risk of data corruption.
For instance, chunks might be cleaned up in the current sqlite
transaction, waiting for the next manifest flush to be committed.
But if a commit occurs before the flushing the current manifest,
the system enters a risky state: the chunk is cleared from the disk
but the current version of the manifest is still in memory. If a
crash happens at this point, the database will stay in its current
state, i.e with a manifest referencing non-existing blocks.

This is solved by adding the IDs of the chunks to cleanup to the
data structure that keeps track of the unwritten manifests. This
way, the cleanup can occur in between the writing of the manifests
and the committing of the transaction.