Force UserStorage init ensure user manifest exists in storage
Unlike other manifest, user manifest is always awaible given we can
guess version 0 (only the `created` field can be considered wrong,
but it will be corrected when merging with remote manifest).
The previous way of lazily generating user manifest v0 create a special
behavior which messed with the realm checkpoint handling code (which
considers a remote modification can be safely ignored if the corresponding
manifest is not in our local database...).
This fix make sure the user manifest is always present in the local
database and insert a generated and version 0 if it's not the case.