Disable reject_irp_prior_to_transact0 option for WinFSP
Enabling reject_irp_prior_to_transact0 is a fix to correct an incompatibily with
Avast when using a folder as mountpoint.
However this fix also prevents from knowing with precision when the mountpoint
is ready which lead to flaky errors in the CI.

Since Parsec have switched to drive based mountpoint (given doing folder-based
is basically a dumpster fire full of nonsenses like the Avast craziness we got
in the first place...) this fix is no longer required.