Get rid of the conflict name when an entry has been renamed both locally and remotely.
This conflict name was unecessary as a better approach is to simply keep the remote name
since it corresponds to a user that managed to synchronize its change before the local
device. Adding a conflict name in this case doesn't help because it'll then be synchronized
and merged on other devices, potentially creating confusion and more conflicts.

There are only two conflict names remaining in the code base:
- Two different files have been given the same name locally and remotely. There is no way
  around it, one of those file have to be renamed to explain this conflict in names.
- A file has been modified both locally and remotely. There is no way around it either,
  there are now two versions of the same file so one of them has to be copied and given a
  new name.