Only use `127.0.0.1` (instead of `localhost`) in tests
The rational is that localhost can resolve to `127.0.0.1` (IPv4) or `::1` (IPv6)
or even any other value if current machine is exotically configured.

On top of that, `trio.open_tcp_stream` may open multiple listeners if `localhost`
resolves to both IPv4 and IPv6.

So forcing to `127.0.0.1` shield us from random behavior that may make the test
unreliable in an unsuspecting way.