Relax `test_eyeballs` expected delay timeout to 1s
On the test `test_eyeballs` we check if it don't take too much time to execute.
The minimal possible value for the test is `800ms` (`TIMEOUT * 2 + 300`).
Previously we were adding 100ms for the overhead of async code but time to time the test will fail.
Here we're adding another 100ms for the overhead and should drastically reduce the number of false positive