Improve `check_newsfragment` check.
When we've multiple newsfragment and only one is valid, the check may failed.
Because it check the one that aren't valid and raise an exception before checking the valid one.

This rework the function that does the check in a manner that don't raise exception on validation error
but will return `False` instead.
On top of that the function will return `None` for older news fragment, thus allowing to filter-out None element to see if the user have added a new newsfragment(s).