Stabilize introduced field in a prior version
Let's say we have introduced some fields in the version `2.3`.
We can add new fields in a major release but those would be unstable since we can't
unsure that all client/server use a version that include those fields, so they're wrapped in `Maybe`.

But when we release a new major version (the version `3.0` from the version `2.3`),
those fields can now be considered stable and don't need the `Maybe` wrapping.