Async local database (#3916)
* Re-export `futures` crate + primitive sync

* Create crate to execute sql query in sub-thread

* Rework `libparsec_core_fs::storage`

- Use new async `LocalDatabase` for `libparsec_core_fs`
- Use async Mutex

* `FutureIntoCoroutine` now accept `Future` that return value

* Update pyo3 storage to use new `libparsec_core_fs::storage` API

* Update python binding with new async storage classes

- Update binding for the new user storage async class
- Update binding for the new workspace storage async class

* Correct test failure when using 2 device at the same time.

The tests `reencrypt` where failing because we would first configure a workspace
shared with `alice` and `bob` with `alice` as admin which is connection will remain open
(meaning a `UserFS` and `UserStorage` instances are running for that device)
and reuse alice do to some operation on workspace.
The issue was that the first device was receiving the event and keping the second one un-change
But during the test we would use the second one to see if it correct receive the event.

Also improve typing for `logged_gui`