Correct `FutureIntoCoroutine` missing with future result
On some specific condition `FutureIntoCoroutine` was able to return the value from another future.

Imagine we have:

```python
await sleep(50)
res = await fetch_number(42)
assert res == 42
```

If the sleep is cancel but the tokio future isn't when this future terminate and try to notify `trio`.
Trio would interprate that signal as it was from `fetch_number`