Rework `SqliteExecutor::close` to by sync
Closing the db connection can be made in a sync manner but
For that we where required to use a `RwLock` that's sync and thus enforcing
The creation of `ExecJob` to have the method `send(job)` to not be async.
That change can delay the closing of `BackgroundSqliteExecutor`
since `ExecJob` also have a copy of the `send_job` channel and the background thread
only stop when all the `send` channels are close on the job queue is empty.