Configure a mirror list for ubuntu APT repo
Sometime the default configured APT proxy in the ubuntu runner image have connection failure.
Since this Repo is a cache-proxy meaning it redirect to the official Canonical repo on cache miss,
We can have failure if the link between the Cache-proxy and the ubuntu repo fail for whatever reason.

To mitigate that I've applied the suggestion in https://github.com/actions/runner-images/issues/7048#issuecomment-1419426054
We basically make `apt` use a mirror list of APT repos with the Cache-proxy has the top priority one.