[testenv] Rework script to allow cleanup
Previous to cleanup our testenv we would have to close the shell that started the env.
But in my case I would get zombie process that run instance of parsec-backend.

I rework the script to have a `cleanup` function that the user can call to kill the backend + remove file + clean env var.