Remove python deps `PyNaCl`
Here we remove the dependency to `PyNaCl`

For that we remplate the exception `nacl.CryptoError` by one defined in the python-rust binding.

We also have to replace the fixture `mock_crypto`.
For that we introduce the feature `test-unsecure-but-fast-secretkey-from-password` in `libparsec_crypto`
to replace the Argon2 algorithm in `SecretKey::from_password` with a very basic algorithm that is unsecure but fast.
This feature MUST ONLY be used in test because it leak the password information.

Other Changes
-------------

- Add function `api_crypto::add_mod` that export Struct & fn to python module
- Also remove fixture `realcrypto` & option `--realcrypto`.
- Replace `uuid4` with `EnrollmentID`