Merge Manifest & Chunk Storage together
`Manifest` & `Chunk` storage can be merge together because they're using the same `database` file + `ManifestStorage` need `ChunkStorage`'s tables.

Since this is the lowest level possible for the abstraction of the storage and is highly dependent on the platform the code is run-on (we can use `sqlite` on desktop but in a web context we don't have that).

I choose to use trait on `platform_storage` because they're simpler to use and will help us to implement the equivalent object in web.

Other changes
-------------

- More picky on error in `get_manifest` before returning `LocalMiss`.
- Factorize `remove_chunks` `ManifestStorage` & `ChunkStorage` both have a way to remove chunks, I've extracted that logic in a specific fonction.
- Remove `Remanence` interface.

Co-authored-by: Emmanuel Leblond <touilleMan@users.noreply.github.com>