Support `--version` flag almost everywhere
Currently, the `--version` is only handled on the root CLI

```
python -m parsec.cli --version
```

But the installed `parsec` app is aliased like that

```
python -m parsec.cli core gui
```

So previously when adding the flag `--version` result in the error described in #4363.

To fix that, I've added `click.version` in the function `debug_config_options`
So everywhere this function is used to add the `--debug` flag, we also get the `--version` flag.

I've added more case for the test `test_version` to show it's working with sub-command

Closes #4363