Don't require to install `poetry` & `python` in Q&A.
Installing `poetry` & `python` in `Q&A` was only required when we change python code.

It's more efficient to do that in `ci-python` since this workflow should be run when actual python code is changed (note, it is also run when we modify `rust` code).

Other Changes
-------------

- Action `use-pre-commit`:
  - Return the install path of `pre-commit` in the outputs variable `install-path`.
  - Skip `Run pre-commit` step when `install-only != true`.

Update .github/actions/use-pre-commit/action.yml

Co-authored-by: Marcos Medrano <786907+mmmarcos@users.noreply.github.com>