Replace workflow `release-starter` with updated `releaser.py`
The workflow `release-starter` could not work since we require to use signed tag and github action don't allow it currently.

I've discussed with @touilleMan and it will be simpler to do what `release-starter` was doing directly in `releaser.py`.

So `releaser.py` will now:

- Check if we are on the correct release branch (will create it for the first version).
- Show the generate `git-tag`.
- Push the release branch and the tag to the remote.

On top of that, since we require the tag to be signed that mean the user can also sign the commit.

When a release is stable (no prerelease part, dev or local) I've added the `acknowledge` sub-command that will create a Pull-request from the release to the main branch.

Other Changes
-------------

- Fix `mypy` error with missing module `version_updater`.

Co-authored-by: Alban Costaz <41511029+Ironicbay@users.noreply.github.com>
Co-authored-by: Marcos Medrano <786907+mmmarcos@users.noreply.github.com>