Rework how we determine if a workflow should be run
The dispatch is now done in 1 job in the workflow `ci.yml` named `dispatch`.

This Simplify workflow `ci-{web,python,rust}` has they don't need to check if they need to run.

This reduce the CI time consumption as the check has to be done only 1 time in the job `dispatch`.

The new workflow diagram for the CI look like:

```mermaid
flowchart LR
    dispatch

    dispatch -- "If it need to run" --> rust
    dispatch -- "If it need to run" --> python
    dispatch -- "If it need to run" --> web

    quality-assurance --> ci-is-happy
    spelling --> ci-is-happy
    dispatch -- "Used to check that dispatched jobs are successful" --> ci-is-happy
    rust --> ci-is-happy
    python --> ci-is-happy
    web --> ci-is-happy
```