Add proc macro to generate protocol tests layout
It behaves in a similar way to existing macros: it crawls JSON schemas
in the specified path and generates test layout for protocol commands.

Files in `/protocol/tests/<family>` contain skeleton test functions for
protocol comands. They are meant to be implemented incrementally later
on (hence the `TODO` comments instead of `unimplemented!` to allow for
a smooth transition during test migration). An example is implemented
for the vlob_read command's req.

Regarding existing proc macro, a `ReuseSchemaStrategy` was added to
`GenCmdsFamily::new` in order to disable schema reuse strategy during
tests generation.