Avoid vendoring openssl when not needed
Depending on packer's needs, we may or may not want to vendor OpenSSL
(e.g. Docker image already ships openssl so no need to vendor, wheel
distributed on pypi requires vendoring to comply with manylinux)
This is only needed for Linux as other OSs don't provide OpenSSL (and
hence it obviously must be vendored anytime we need it).