Make idempotent client/workspace start in the bindings
We now accept that functions involving client/workspace handle might
return an internal error due to a concurrent stop, and hence that the
GUI has to act accordingly.
The downside of this approach is it is harder to detect if the invalid
handle use is due to a bug, but we accept this given it lower the overall
complexity of the GUI (can be stateless).