Changelog: linked issue's link is on a new line
When generating a news element for `history.rst` we added the link at the end of the text and then wrap-it.
That could result sometime on the link being split.

The link being split don't cause problem once the document is rendered but it's nice to have a functioning link when viewing the raw document.