Backport 2.16 documentation modification (#5653)
* Add command `watch-html` to build the html while editing it.

* Add documentation to install the parsec backend on-premise

This documentation provide a basic `parsec` infra to test the parsec
solution.
It rely on `docker-compose` to host the needed services for the infra
like the database, the s3 and the server.

Co-authored-by: Marcos Medrano <786907+mmmarcos@users.noreply.github.com>

* Replace `docker-compose` with proper one in `docs/adminguide/hosting`

The `docker-compose` at the root of the repository is replaced by
`docs/adminguide/hosting/parsec-server.docker.yaml`

* Fix docs docker-image build related to change introduced in #5573

* Remove `development.po`

This translation file was referencing an old `development.rst` file that no longer exist
(it was remplaced with the actual `development` folder that is no longer included in the documentation).

* Delete `functional_architecture.po`

That file provided the french translation for the file `functional_architecture.rst` that was migrated to
`development/functional_architecture.md` and is not longer translated.

* Changelog: linked issue's link is on a new line

When generating a news element for `history.rst` we added the link at the end of the text and then wrap-it.
That could result sometime on the link being split.

The link being split don't cause problem once the document is rendered but it's nice to have a functioning link when viewing the raw document.

* Add poutils

- Add `poutils` in the pyproject dependencies to provide `powrap.
- Add `gettext` in the dockerfile as `powrap` require `msgcat` tool.
- Add makefile rule `intl-%` to update the translation in the language `%`.
  This rule will also use `powrap` to format the `.po` file that were updated.
- Format PO file using `powrap`

* Configure `gettext` to generate a `pot` file for each `rst` file.

* Split `adminguide.po` into a po files per pages.

* Split `userguide.po` into a po file per page

* Add workflow to check the doc build status

- Update `version_updater` to handle `ci-docs.yml`.
- Update makefile to allow to override the variables:
  - `SPHINXOPTS`
  - `SPHINXBUILD`
  - `PAPER`
  - `BUILDDIR`
- Update `ci.yml` to dispatch the job for the docs.

Co-authored-by: Marcos Medrano <786907+mmmarcos@users.noreply.github.com>

* Simplify documentation versions

* Handle `docs/pyproject.toml` python versions

---------

Co-authored-by: Marcos Medrano <786907+mmmarcos@users.noreply.github.com>