Prevent concurrent run of `publish` workflow
When publishing a release (publish a release that was on draft), that
raise 2 events `published` & `edited`.

We configured the workflow to run when one of those is received but
since publishing a release raise those 2 events, that make the workflow
run 2 time instead of one.

By adding the `concurrent` settings, we prevent the workflow to run on
the same `ref`.