Fix releaser schedule cannot create release because of missing tag.
Previously we didn't have that problem since that step was run when we push a tag and
the action was able to determine the tag name from the event info.

Additional, the step `Remove previous nightly release` was failing because of the missing env var `GH_TOKEN`.

`permissions.contents: write` was added to be more explicit on the required permissions.