Support nightly & release-candidate channel
The previous implementation only support for alpha & beta channel.
We add 2 new channels: nightly & release-candidate.
We also support for aliasing the channel name, for example, `a` for alpha, `b` for beta, `rc` for release-candidate.

By providing the env variable `FORCE_NIGHTLY=1`, the updater will force the channel to be nightly.

We added a new option `--nightly` to `package.js` to allow the user to specify the current build is a nightly build.