Document a quirk in testbed platform storage behavior
Testbed platform storage always saves all certificates in the storage,
this is convenient but doesn't correspond to the actual in-production
behavior where a given user has access of a subset of all realm & shamir
certificates.

In practice this means the higher layers (i.e. `libparsec_client`) are
considered to be aware of this and to implement their code accordingly
(i.e. not crawling over all certificates whithout proper filtering).