Improve cspell configuration
The main improvement is to avoid using the `fr-fr` dictionary for all
files (which led to a lot of typos!). Now, the `overrides` option is
used to use the `fr-fr` dictionary only for files that are supposed to
contain French words.

Other changes:

- Added dictionaries: `docker`, `en-common-misspellings`, `fonts`, `html`, `latex`, `makefile`, `npm`, `powershell`, `shellscript`
- Added `flagword` for "occurred" misspellings (until `dict-cpp` fix is included in `cspell-cli`)
- Removed `words` and `ignoreWords` from `cspell.config.yml` (or moved them to custom dictionaries)
- Removed custom dictionaries: `jinja`, `french-words`
- Removed unnecessary imports

Bonus: unused words removed from custom dictionaries (see
https://github.com/streetsidesoftware/cspell/issues/808#issuecomment-1003374305)