ci(releaser): Fix missing if cond for dependent workflows
With PR-#10159 the if condition that outdated and the job releaser thus
never run because of the condition expecting the result from the
splitted job using the previous combined `job_id`.

Uppon fix that I've figured out that that check for
`package-parsec-server` was invalid (always true, were we want to check
if the job succeeded) and we were missing the `package-parsec-cli`
check.

Fix #10164