refactor(device-loader): Use `StorageManager` over `LocalStorage` on web
Since `libparsec` is run from a shared worker, we no longer have access
to `localStorage` via the `window` property.

This commit refactor the web impl to use `StorageManager` over
`LocalStorage` that is available in both browser and shared worker.

Another fix would have been to store the device in IndexedDB.

- Require to use patch `web-sys` to be able to use the missing method
  `Blob.bytes`.