Remove `/testbed/account/new` route from testbed server
The route is not used (there is not account-related test on client side yet), and
it cumbersome since there is no easy discriminant for concurrent client tests against
a single testbed server (like what we use the organization ID for in the rest of the tests).

So better wait and figure out if it is really needed, typically we may want instead to expose
the sent mails through a `/testbed/mailbox/get` API in order for the client to obtain the
validation token.