Move `PasswordAlgorithm` into `libparsec_crypto` and provides it with dedicated methods
The key need here is to keep together the code responsible for creating a new password algorithm (which
relies on some constants for the parameters values, e.g. `opslimit`) and the one for faking a config
for a non-existing account.
This is important to avoid both from diverging when new algorithm and config are added.